Create Trophy Case Candidates.md
A list of games I  could possibly consider worthy of completing and adding to my Trophy Case in my Trophy Room scene on stream.

These can be any game I feel like commemorating, but an easier target to hit is games that are very difficult but not impossible. Ideally a game it will take me a significant amount of practice to finish.

Something akin to the process of finishing your first Dark Souls game (I know, it's cliche, but I don't have a better short hand for it).